ipcache map: extend cluster_id to uint16
Extend the cluster_id field of the current ipcache map key (struct
ipcache_key/Key) from uint8 to uint16 to support larger Clustermesh.

+++ pkg/datapath/ipcache/listener.go
@@ -143,7 +143,7 @@ func (l *BPFListener) OnIPIdentityCacheChange(modType ipcache.CacheModification,
 
 	// Update BPF Maps.
 
-	key := ipcacheMap.NewKey(cidr.IP, cidr.Mask, uint8(cidrCluster.ClusterID()))
+	key := ipcacheMap.NewKey(cidr.IP, cidr.Mask, uint16(cidrCluster.ClusterID()))
 
 	switch modType {
 	case ipcache.Upsert:
